% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_target_score.R
\name{calc_target_score}
\alias{calc_target_score}
\title{Compute Target score for randomized data/compute P value for each TS given a network topology}
\usage{
calc_target_score(
  wk,
  wks,
  dist_ind,
  inter,
  n_dose,
  n_prot,
  proteomic_responses,
  fs_dat,
  verbose = TRUE,
  ts_factor = 1,
  dist_file = NULL
)
}
\arguments{
\item{wk}{Inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate and -1 for down regulate.}

\item{wks}{Inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate, -1 for down regulate,
2 for phosphorylation and -2 for dephosphorylation.}

\item{dist_ind}{A distance file of edgelist with a third column as the network distance
between the genes in the interaction}

\item{inter}{Edgelist of inferred network}

\item{n_dose}{Dose number of input data.}

\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{fs_dat}{Functional score file. A tab-delmited file with a header, each row is an
antibody in the first column and functional score in the second column
(i.e. 1 oncogene, 0 tumor supressor/oncogene, -1 tumor supressor characteristics)}

\item{verbose}{a boolean to show debugging information}

\item{ts_factor}{a scaling factor for the pathway component in the target score}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}
}
\description{
Compute Target score for randomized data/compute P value for each TS given a network topology
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
missing: For phosp and dephosp based wk, there is no 'exact match' between known and measured phospho-sites
}
\concept{zeptosensPkg}
