% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIntermediateFormatToMatrix.R
\name{convertIntermediateFormatToMatrix}
\alias{convertIntermediateFormatToMatrix}
\title{Convert Intermediate Format to Matrix}
\usage{
convertIntermediateFormatToMatrix(arrayData, colsForRowNames)
}
\arguments{
\item{arrayData}{a data.frame as described in readZeptosensExport}

\item{colsForRowNames}{a vector of column numbers to be used for creating row names}
}
\value{
a matrix with antibodies as the column names and row names using the user selected columns 
  separated by underscores
}
\description{
Convert Intermediate Format to Matrix
}
\concept{
zeptosensPkg
}

