% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sif_from_matrix.R
\name{create_sif_from_matrix}
\alias{create_sif_from_matrix}
\title{Create sif from network matrix.}
\usage{
create_sif_from_matrix(
  t_net,
  row_genelist = rownames(t_net),
  col_genelist = colnames(t_net)
)
}
\arguments{
\item{t_net}{p*p Gaussian gene Network matrix estimated.}

\item{row_genelist}{Gene name list corresponding to the row of gene Network matrix}

\item{col_genelist}{Gene name list corresponding to the column of gene Network matrix.}
}
\value{
Edgelist description data.frame contain G(V,E) including Vertex and Edges
(1 positive/-1 negative) edgevalue (the strength of association) and vertices numbers from the graph
}
\description{
Create sif from network matrix.
}
\examples{
network_org <- readRDS(system.file("test_data_files", "predict_bio_network_network_output.rds",
package = "targetscore"
))
edgelist_wk <- create_sif_from_matrix(
t_net = network_org$wk,
col_genelist = colnames(network_org$wk),
row_genelist = rownames(network_org$wk)
)

}
\concept{targetscore}
