% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getChibeExport.R
\name{genChibeExport}
\alias{genChibeExport}
\title{Generate ChiBE Export}
\usage{
genChibeExport(antibodies, antibodyMapFilename = NULL,
  chibeFilename = "abChibe.txt")
}
\arguments{
\item{antibodies}{FIXME}

\item{antibodyMap}{a string, antibody map file name (default: NULL, if NULL then use zentosensPkg internal antibody map)}
}
\description{
Generate ChiBE Export
}
\concept{
zeptosensPkg
}

