% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetScore.R
\name{getTargetScore}
\alias{getTargetScore}
\title{Target Score Pilot Run}
\usage{
getTargetScore(nDose, nProt, proteomicResponses, maxDist = 1, nPerm, cellLine,
  targetScoreOutputFile = NULL, matrixWkOutputFile = NULL,
  targetScoreQValueFile = NULL, targetScoreDoseFile = NULL)
}
\arguments{
\item{nDose}{TBA}

\item{nProt}{TBA}

\item{proteomicResponses}{TBA}

\item{maxDist}{TBA (default: 1)}

\item{nPerm}{number of random TS calculations for building the null distribution}

\item{cellLine}{TBA}

\item{targetScoreOutputFile}{a filename to write total target score results (default: NULL)}

\item{matrixWkOutputFile}{TBA}

\item{targetScoreQValueFile}{a file namme to write statistical significance levels (default: NULL)}

\item{targetScoreDoseFile}{a filename to write dose dependent target score results (default: NULL)}
}
\description{
Target Score Pilot Run
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
missing: For phosp and dephosp based wk, there is no "exact match" between known and measured phospho-sites
}
\examples{

}
\concept{
zeptosensPkg
}

