% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_genes_to_edgelist.R
\name{match_genes_to_edgelist}
\alias{match_genes_to_edgelist}
\title{Match a vector of genes to a SIF network}
\usage{
match_genes_to_edgelist(
  genes1,
  genes2 = NULL,
  annot_edgelist,
  antibody_vec,
  use_annot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{genes1}{a vector of genes}

\item{genes2}{an optional vector if the source}

\item{annot_edgelist}{a data.frame; the first two columns are interaction
participants and the third column is an optional annotation}

\item{antibody_vec}{a vector with the names of the antibodies (e.g. colnames(proteomicResponses));
indicies returned from this function will be mapped to this vector}

\item{use_annot}{a boolean as to whether the optional annotation column values
in the annot_edgelist should outputted in the results}

\item{verbose}{a boolean to show debugging information}
}
\value{
a data.frame, columns 1-2 are indicies of the edgelist participants,
  column 3 is the annotation values for the edgelist, 4-5 the names of the
  edgelist participants
}
\description{
Match a vector of genes to a SIF network
}
\examples{
antibody_map_file <- system.file("targetScoreData", "antibodyMapfile_08272020.csv", 
  package = "targetscore")
mab_to_genes <- read.csv(antibody_map_file, header = TRUE, stringsAsFactors = FALSE)
proteomic_responses_file <- system.file("test_data", "BT474.csv", package = "targetscore")
proteomic_responses <- read.csv(proteomic_responses_file, row.names = 1)
dist_file <- system.file("targetScoreData", "distances.txt", package = "targetscore")
tmp_dist <- read.table(dist_file, sep = "\t", header = TRUE, stringsAsFactors = FALSE)

dist <- tmp_dist[which(tmp_dist[, 3] <= 1), ]
idx_ab_map <- which(mab_to_genes[, 1] \%in\% colnames(proteomic_responses))

mab_genes <- mab_to_genes[idx_ab_map, 4]
names(mab_genes) <- mab_to_genes[idx_ab_map, 1] 

dist_list <- match_genes_to_edgelist(
genes1 = mab_genes,
genes2 = NULL,
annot_edgelist = dist,
antibody_vec = colnames(proteomic_responses),
use_annot = TRUE,
verbose = TRUE
)

}
\concept{targetscore}
