% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameter_hybrid.R
\name{optimize_parameter_hybrid}
\alias{optimize_parameter_hybrid}
\title{Choose the optimal regulization parameter and scale paramter for adjusted-glasso algorithm network construction}
\usage{
optimize_parameter_hybrid(
  data,
  prior = NULL,
  rho = 10^seq(-2, 0, 0.02),
  kappa = 10^seq(-2, 0, 0.02),
  antibody_map_file = system.file("targetscoreData", "antibody_map_08092019.csv",
    package = "targetscore")
)
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{prior}{Prior information matrix of gene interaction, with colnames and rownames as gene tags.
With colnames and rownames as gene tags.Can be inferred from Public data source (for example:SignedPC).}

\item{rho}{positive tuning parameter for elastic net penalty. Default at seq10^seq(-2, 0, 0.02).}

\item{kappa}{positive scaler parameter for biology-knowledge base contribution. Default at seq10^seq(-2, 0, 0.02).}

\item{antibody_map_file}{a CSV file containing a table columns: 
AntibodyLabel (a unique for antibody), Source (supplier), NodeName (a label for entry for visualization), 
Gene_Symbol (approved gene symbol) Sites (affected sites divided by vertical bar, |), 
Effect (effect of phosphorylation: c: constitutively expressed, a: activating, i: inhibiting)}
}
\value{
Parameter list of regulization parameter decided by the prior information and the algorithmn lowest BIC.
\item{rho}{optimized penalty parameter.}
\item{kappa}{optimized scaler parameter.}
\item{rho_m}{optimized regulization matrix for the expression data.}
\item{bic}{Calculated model BIC error for differnet regularization parameters.}
}
\description{
Choose the optimal regulization parameter and scale paramter for adjusted-glasso algorithm network construction
}
\examples{
# read proteomic response file
file <- system.file("test_data", "TCGA-BRCA-L4.csv", package = "targetscore")
signaling_responses <- read.csv(file, row.names = 1)

# Read in Biology knowlegde base protein interaction
file <- system.file("test_data_files", "predict_bio_network_network_output.rds",
  package = "targetscore")

prior_org <- readRDS(file)
parameters <- targetscore::optimize_parameter_hybrid(data = signaling_responses, 
  prior = prior_org$wk)

}
\concept{targetscore}
