% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randTargetScore.R
\name{randTargetScore}
\alias{randTargetScore}
\title{Compute Target score for randomized data/compute P value for each TS given a network topology}
\usage{
randTargetScore(nDose, nProt, proteomicResponses, maxDist = 1, cellLine)
}
\arguments{
\item{nDose}{TBA}

\item{nProt}{TBA}

\item{proteomicResponses}{TBA}

\item{maxDist}{TBA (default: 1)}

\item{cellLine}{TBA}

\item{targetScoreOutputFile}{a filename to write target score results (default: NULL)}

\item{matrixWkOutputFile}{TBA}
}
\description{
Compute Target score for randomized data/compute P value for each TS given a network topology
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
missing: For phosp and dephosp based wk, there is no "exact match" between known and measured phospho-sites
}
\examples{

}
\concept{
zeptosensPkg
}

