% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readZeptosensExport2.R
\name{readZeptosensExport2}
\alias{readZeptosensExport2}
\title{Read Zeptosens Export File}
\usage{
readZeptosensExport2(data, sampleNameEntries = NULL)
}
\arguments{
\item{sampleNameEntries}{TBA}

\item{inputFile}{TBA}
}
\description{
Read Zeptosens Export File
}
\examples{
tmp <- readZeptosensExport("inst/dataInst/R006_RFI_Export_Table.xls", c("sampleNameNumber", "NA", "NA", "sampleName", "treatment", "time", "replicate"))
}
\concept{
zeptosensPkg
}

