% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robustness_test.R
\name{run_robustness_test}
\alias{run_robustness_test}
\title{Run robustness test for data through the chosen algorithm.}
\usage{
run_robustness_test(
  data,
  prior = NULL,
  proteomic_responses,
  n_prot,
  max_dist = 1,
  boot_time = 100,
  cut_off = 0.1,
  mab_to_genes,
  algorithm = c("data_driven", "hybrid_driven")
)
}
\arguments{
\item{data}{input proteomics expression dataset for network inference. Coloumns as the gene, rows as the sample.With
colnames as the gene tags, rownames as the sample tags.}

\item{prior}{Prior information data frame ,with colnames and rownames as gene tags.
With colnames and rownames as gene tags. Can be inferred from predict_bio_network() or any network resources.
Default at NULL and will be inferred from SignedPC database.}

\item{proteomic_responses}{RPPA data tested for drug pertubation.}

\item{n_prot}{Antibody number of input data.Here should be the number of antibody exist both in network
inference dataset and proteomic_responses dataset.}

\item{max_dist}{maximum distance between two antibody. (Default at 1)}

\item{boot_time}{Bootstrap time mannually set.Default at 100.}

\item{cut_off}{Manually set up cut off value for strength of edge. (Default at 0.1)}

\item{mab_to_genes}{A list of antibodies, their associated genes, modification sites and effect.}

\item{algorithm}{Flexible toolbox implementing network estimating algorithms for robustness test.
(\code{data_driven},or \code{hybrid_driven}).}
}
\value{
result of validation score. for random network and the network predicted with the algorithm.
}
\description{
Run robustness test for data through the chosen algorithm.
}
\concept{targetscore}
