% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unperturbednormalization.R
\name{unperturbednormalization}
\alias{unperturbednormalization}
\title{Normalize Zeptosens Data wrt unperturbed (or any other condition)}
\usage{
unperturbednormalization(Array1, normal_factor, nsamples, antibodyNum)
}
\arguments{
\item{normal_factor}{(treatment_name)}

\item{nsamples}{samples/Ab}

\item{antibodyNum}{number of antibodies on array}

\item{array1}{FIXME}
}
\description{
Normalize Zeptosens Data wrt unperturbed (or any other condition)
}
\details{
Normalize wrt ttl prt lvl
}
\concept{
zeptosensPkg
}

