% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dat_network.R
\name{predict_dat_network}
\alias{predict_dat_network}
\title{predict data-driven only network.}
\usage{
predict_dat_network(
  data,
  cut_off = 0.1,
  n_prot,
  proteomic_responses,
  rho = 10^seq(-2, 0, 0.02)
)
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{cut_off}{Manually Setup cut off value for the strength of edge. Default at 0.1.}

\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{rho}{positive tuning parameter vector for elastic net penalty. Default at 10^seq(-2,0, 0.02).}
}
\value{
Parameter of regulization decided lowest BIC.Including regularize parameter(L1 norm parameter) as rho.
}
\description{
predict data-driven only network.
}
\examples{
# Read proteomic response for cellline1
file <- system.file("test_data", "BT474.csv", package = "zeptosensPkg")
proteomic_responses <- read.csv(file, row.names = 1)

# Read Global Signaling file for BRCA
file <- system.file("test_data", "TCGA-BRCA-L4.csv", package = "zeptosensPkg")
signaling_responses <- read.csv(file, row.names = 1)

 # Extract network
 network <- zeptosensPkg::predict_dat_network(
 data <- signaling_responses,
 n_prot = dim(proteomic_responses)[2],
 proteomic_responses = proteomic_responses
 )

}
\concept{zeptosensPkg}
