% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readZeptosensExport.R
\name{readZeptosensExport}
\alias{readZeptosensExport}
\title{Read Zeptosens Export File}
\usage{
readZeptosensExport(inputFile = NULL, sampleNameEntries = NULL)
}
\arguments{
\item{inputFile}{TBA}

\item{sampleNameEntries}{TBA}
}
\description{
Read Zeptosens Export File
}
\examples{
tmp <- readZeptosensExport("inst/dataInst/R006_RFI_Export_Table.xls", c("sampleNameNumber", "NA", "NA", "sampleName", "treatment", "time", "replicate"))
}
\concept{
zeptosensPkg
}

