% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readZeptosensExport.R
\name{readZeptosensExport}
\alias{readZeptosensExport}
\title{Read Zeptosens Export File}
\usage{
readZeptosensExport(data, sampleNameEntries = NULL)
}
\arguments{
\item{data}{Zeptosens data read in by readZeptosensXls}

\item{sampleNameEntries}{FIXME}
}
\description{
Read Zeptosens Export File
}
\details{
Format: sampleNumber_sample_treatment_dose_time_replicate_date_notes
}
\examples{
tmp <- readZeptosensExport(tmp, c("sampleNumber", "NA", "NA", "sample", "treatment", "time", "replicate"))
}
\concept{
zeptosensPkg
}

